refactor: move appEvents file
