+++ imports/plugins/core/cart/server/methods/setUserCurrency.js
@@ -3,7 +3,7 @@ import ReactionError from "@reactioncommerce/reaction-error";
 import { check, Match } from "meteor/check";
 import { Cart } from "/lib/collections";
 import getCart from "/imports/plugins/core/cart/server/util/getCart";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 
 /**
  * @method cart/setUserCurrency
