+++ imports/plugins/core/cart/server/methods/submitPayment.js
@@ -5,7 +5,7 @@ import { Cart } from "/lib/collections";
 import Reaction from "/imports/plugins/core/core/server/Reaction";
 import ReactionError from "@reactioncommerce/reaction-error";
 import getCart from "/imports/plugins/core/cart/server/util/getCart";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 import { PaymentMethodArgument } from "/lib/collections/schemas";
 
 /**
