+++ imports/plugins/core/cart/server/methods/unsetAddresses.js
@@ -4,7 +4,7 @@ import { check, Match } from "meteor/check";
 import ReactionError from "@reactioncommerce/reaction-error";
 import { Cart } from "/lib/collections";
 import getCart from "/imports/plugins/core/cart/server/util/getCart";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 
 /**
  * @method removeShippingAddresses
