+++ imports/plugins/core/cart/server/no-meteor/mutations/convertAnonymousCartToNewAccountCart.js
@@ -1,7 +1,7 @@
 import Random from "@reactioncommerce/random";
 import ReactionError from "@reactioncommerce/reaction-error";
 import { Cart as CartSchema } from "/imports/collections/schemas";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 
 /**
  * @summary Copy items from an anonymous cart into a new account cart, and then delete the
