+++ imports/plugins/core/discounts/server/hooks/cart.js
@@ -1,6 +1,6 @@
 import Hooks from "@reactioncommerce/hooks";
 import { Meteor } from "meteor/meteor";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 import { Cart } from "/lib/collections";
 
 appEvents.on("afterCartUpdate", (cartId, cart) => {
