+++ imports/plugins/core/discounts/server/methods/methods.js
@@ -2,7 +2,7 @@ import { Meteor } from "meteor/meteor";
 import { Match, check } from "meteor/check";
 import ReactionError from "@reactioncommerce/reaction-error";
 import { Cart } from "/lib/collections";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 import { Discounts } from "../../lib/collections";
 import Reaction from "../api";
