+++ imports/plugins/core/graphql/server/index.js
@@ -1,6 +1,6 @@
 import { Meteor } from "meteor/meteor";
 import { WebApp } from "meteor/webapp";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 import collections from "/imports/collections/rawCollections";
 import createApolloServer from "/imports/node-app/core/createApolloServer";
 import runPluginStartup from "./no-meteor/runPluginStartup";
