+++ imports/plugins/core/payments/server/methods/apply.js
@@ -1,7 +1,7 @@
 import Random from "@reactioncommerce/random";
 import { check } from "meteor/check";
 import Reaction from "/imports/plugins/core/core/server/Reaction";
-import appEvents from "/imports/plugins/core/core/server/appEvents";
+import appEvents from "/imports/node-app/core/util/appEvents";
 
 /**
  * @name payments/apply
