+++ tests/TestApp.js
@@ -3,7 +3,7 @@ import graphql from "graphql.js";
 import findFreePort from "find-free-port";
 import MongoDBMemoryServer from "mongodb-memory-server";
 import Random from "@reactioncommerce/random";
-import appEvents from "../imports/plugins/core/core/server/appEvents";
+import appEvents from "../imports/node-app/core/util/appEvents";
 import createApolloServer from "../imports/node-app/core/createApolloServer";
 import defineCollections from "../imports/node-app/core/util/defineCollections";
 import Factory from "../imports/test-utils/helpers/factory";
